requirements.txt
alembic==1.0.11
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.12
MarkupSafe==1.1.1
PyJWT==1.7.1
PySocks==1.7.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.1
requests==2.22.0
schedule==0.6.0
six==1.12.0
SQLAlchemy==1.3.5
twilio==6.28.0
urllib3==1.25.3
Werkzeug==0.15.4
WTForms==2.2.1
